refactor: remove redundancy from sidebar components, re-order import in components
