+++ frontend/src/components/server/ServerLayout.tsx
@@ -1,16 +1,14 @@
-import React, { FunctionComponent, useState } from 'react'
+import React, { FunctionComponent } from 'react'
 import { useAppSelector } from 'redux/hooks'
 
-import ServerSidebar from './sidebar/ServerSidebar'
-import ServerNavbar from './navbar/ServerNavbar'
-import AddChannelModal from 'components/server/modal/AddChannelModal'
-import MetaPanel from 'components/server/metaPanel/MetaPanel'
-import ServerMessages from 'components/server/messagePanel/ServerMessages'
 import { selectIsMetaPanelOpen } from 'components/server/redux/metaPanel.slice'
 
-const ServerLayout: FunctionComponent = () => {
-    const [isChannelModalOpen, setChannelModalOpen] = useState<boolean>(false)
+import ServerNavbar from './navbar/ServerNavbar'
+import ServerSidebar from './sidebar/ServerSidebar'
+import ServerMessages from './messagePanel/ServerMessages'
+import MetaPanel from './metaPanel/MetaPanel'
 
+const ServerLayout: FunctionComponent = () => {
     const isMetaPanelOpen = useAppSelector(selectIsMetaPanelOpen)
 
     return (
@@ -19,9 +17,7 @@ const ServerLayout: FunctionComponent = () => {
                 <ServerNavbar />
                 <div className="grid grid-cols-12 w-full h-full">
                     <div className="col-span-2 w-full h-full bg-slack-sidebar-normal">
-                        <ServerSidebar
-                            setChannelModalOpen={setChannelModalOpen}
-                        />
+                        <ServerSidebar />
                     </div>
                     <div
                         className={`${
@@ -37,10 +33,6 @@ const ServerLayout: FunctionComponent = () => {
                         <MetaPanel />
                     </div>
                 </div>
-                <AddChannelModal
-                    isOpen={isChannelModalOpen}
-                    setOpen={setChannelModalOpen}
-                />
             </div>
         </div>
     )
